update and move command
