+++ packages/cli/src/commands/setup/docker/templates/Dockerfile
