+++ packages/cli/src/commands/setup/docker/templates/docker-compose.dev.yml
