+++ packages/cli/src/commands/setup/docker/templates/docker-compose.prod.yml
